pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cartpole"
version = "0.1.0"
description = "Cart-pole control environment with multiple controllers"
authors = [{name="Your Name"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = ["numpy", "scipy", "matplotlib", "pyyaml"]
